Organize files in k2 module
