+++ compiler/k2/src/main/core/com/kitakkun/kondition/compiler/k2/diagnostics/KonditionErrorMessages.kt
@@ -1,4 +1,4 @@
-package com.kitakkun.kondition.compiler.k2
+package com.kitakkun.kondition.compiler.k2.diagnostics
 
 import org.jetbrains.kotlin.diagnostics.KtDiagnosticFactoryToRendererMap
 import org.jetbrains.kotlin.diagnostics.KtDiagnosticRenderers.TO_STRING
