+++ compiler/k2/src/main/core/com/kitakkun/kondition/compiler/k2/diagnostics/KonditionErrors.kt
@@ -1,4 +1,4 @@
-package com.kitakkun.kondition.compiler.k2
+package com.kitakkun.kondition.compiler.k2.diagnostics
 
 import org.jetbrains.kotlin.com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.diagnostics.error1
