+++ compiler/k2/src/main/core/com/kitakkun/kondition/compiler/k2/extensions/KonditionFirExtensionRegistrar.kt
@@ -1,5 +1,7 @@
-package com.kitakkun.kondition.compiler.k2
+package com.kitakkun.kondition.compiler.k2.extensions
 
+import com.kitakkun.kondition.compiler.k2.api.VersionSpecificAPI
+import com.kitakkun.kondition.compiler.k2.api.VersionSpecificAPIImpl
 import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrar
 
 class KonditionFirExtensionRegistrar : FirExtensionRegistrar() {
