+++ compiler/k2/src/main/pre_2_0_10/com/kitakkun/kondition/compiler/k2/api/VersionSpecificAPIImpl.kt
@@ -1,4 +1,4 @@
-package com.kitakkun.kondition.compiler.k2
+package com.kitakkun.kondition.compiler.k2.api
 
 import org.jetbrains.kotlin.fir.FirSession
 import org.jetbrains.kotlin.fir.declarations.evaluateAs
@@ -7,6 +7,6 @@ import org.jetbrains.kotlin.fir.expressions.FirLiteralExpression
 
 object VersionSpecificAPIImpl : VersionSpecificAPI {
     override fun evaluateAsLiteralValue(originalExpression: FirExpression, session: FirSession): Any? {
-        return originalExpression.evaluateAs<FirLiteralExpression>(session)?.value
+        return originalExpression.evaluateAs<FirLiteralExpression<*>>(session)?.value
     }
 }
